+++ library/src/main/jni/android_local_participant_observer.cpp
@@ -87,7 +87,7 @@ AndroidLocalParticipantObserver::AndroidLocalParticipantObserver(JNIEnv *env,
                   webrtc_jni::GetMethodID(env,
                                           *j_twilio_exception_class_,
                                           "<init>",
-                                          kTwilioExceptionConstructoSignature)){
+                                          kTwilioExceptionConstructorSignature)){
     VIDEO_ANDROID_LOG(twilio::video::LogModule::kPlatform,
                       twilio::video::LogLevel::kDebug,
                       "AndroidLocalParticipantObserver");